+++ src/Game/Text/de/HelpText.de.sjson.lua
@@ -991,7 +991,7 @@ local newData = {
 	-- Add Pool of Purging/SellShops/SellTraitShops to post-boss rooms
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_UnlockPostBossSellShopsIncantation",
-	-- 	DisplayName = "Summoning of Blood-Red Fountains",
+	-- 	DisplayName = "Surge of Blood-Red Fountains",
 	-- 	Description =
 	-- 	"Cause a {$Keywords.SellTraitShop} to always appear in the space between each {$Keywords.Biome} in the {$Keywords.ModsNikkelMHadesBiomesModdedRoute}."
 	-- },
@@ -1011,6 +1011,30 @@ local newData = {
 	-- 	Id = "ModsNikkelMHadesBiomes_UnlockPostBossSellShopsIncantation_Flavor",
 	-- 	Description =
 	-- 	"One would think that any Olympian boon is one to be cherished, but sometimes even the Princess of the Underworld finds herself in need of coin."
+	-- },
+	-- Add Well of Charon/WellShops to post-boss rooms
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_UnlockPostBossWellShopsIncantation",
+	-- 	DisplayName = "Surge of Lilac-Flowing Fountains",
+	-- 	Description =
+	-- 	"Cause a {$Keywords.WellShop} to always appear in the space between each {$Keywords.Biome} in the {$Keywords.ModsNikkelMHadesBiomesModdedRoute}."
+	-- },
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_UnlockPostBossWellShopsIncantation_Flavor",
+	-- 	Description =
+	-- 	"The great Lord Charon has long been known to offer more than just passage across the River Styx - for those who can afford it."
+	-- },
+	-- Add Well of Charon/WellShops to any rooms during a run
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_UnlockInRunWellShopsIncantation",
+	-- 	DisplayName = "Rise of Lilac-Flowing Fountains",
+	-- 	Description =
+	-- 	"Cause {$Keywords.RoomPlural} in the {$Keywords.ModsNikkelMHadesBiomesModdedRoute} to sometimes contain a {$Keywords.WellShop}."
+	-- },
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_UnlockInRunWellShopsIncantation_Flavor",
+	-- 	Description =
+	-- 	"The boatman Charon was of great help to the Prince Zagreus on his attempts to escape the Underworld, always willing to lend a hand - for the right price."
 	-- },
 	-- #endregion