+++ src/Scripts/RoomDataAsphodel.lua
@@ -3,11 +3,18 @@ local roomReplacements = {
 	-- GENERIC
 	BaseAsphodel = {
 		SecretDoorRequirements = game.DeepCopyTable(game.RoomSetData.Base.BaseRoom.SecretDoorRequirements),
-		WellShopRequirements = game.DeepCopyTable(game.RoomSetData.Base.BaseRoom.WellShopRequirements),
 		-- The Asphodel teleport in Hades II - we don't want it in Zagreus' Journey
 		AnomalyDoorChance = 0.0,
 		RoomSetName = "Asphodel",
 
+		WellShopRequirements = {
+			{
+				PathTrue = { "GameState", "WorldUpgrades", "ModsNikkelMHadesBiomes_UnlockInRunWellShopsIncantation" },
+			},
+			RequiredMinCompletedRuns = 1,
+			RequiredMinRoomsSinceWellShop = 3,
+			RequiredMinBiomeDepth = 4,
+		},
 		SellShopRequirements = {
 			{
 				PathTrue = { "GameState", "WorldUpgrades", "ModsNikkelMHadesBiomes_UnlockInRunSellShopsIncantation" },
@@ -581,16 +588,19 @@ local roomModifications = {
 		ThreadedEvents = {
 			[1] = { FunctionName = "HadesSpeakingPresentation", Args = { VoiceLines = game.GlobalVoiceLines.HadesPostBossVoiceLines, StartDelay = 2.5 } },
 		},
-		SellShopRequirements = {
-			{
-				PathTrue = { "GameState", "WorldUpgrades", "ModsNikkelMHadesBiomes_UnlockPostBossSellShopsIncantation" },
-			},
-		},
+		-- In vanilla the game adds a SetupGameStateRequirement to the obstacle, but we don't do that in modded rooms
+		-- So to prevent the well from spawning even without the incantation, we set ForceWellShop to false and just use the WellShopSpawnChance
+		ForceWellShop = false,
 		-- Defines where the well spawns, making sure the sell shop spawns on the other possible ID - available IDs are 480768 and 532755
 		WellShopChallengeBaseId = 480768,
 		WellShopRequirements = {
 			{
-				PathTrue = { "GameState", "WorldUpgradesAdded", "WorldUpgradePostBossWellShops" },
+				PathTrue = { "GameState", "WorldUpgradesAdded", "ModsNikkelMHadesBiomes_UnlockPostBossWellShopsIncantation" },
+			},
+		},
+		SellShopRequirements = {
+			{
+				PathTrue = { "GameState", "WorldUpgrades", "ModsNikkelMHadesBiomes_UnlockPostBossSellShopsIncantation" },
 			},
 		},
 		ObstacleData = {